requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
Babel==2.5.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cliff==2.9.1
cloudflare==1.8.1
cloudflare-python==1.0.1
cmd2==0.7.8
cryptography==2.1.4
debtcollector==1.19.0
deprecation==1.0.1
future==0.16.0
idna==2.6
iso8601==0.1.12
jsonpatch==1.21
jsonpointer==1.14
jsonschema==2.6.0
keystoneauth1==3.3.0
monotonic==1.4
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.6
openstacksdk==0.9.19
os-client-config==1.28.0
osc-lib==1.8.0
oslo.config==5.1.0
oslo.i18n==3.19.0
oslo.serialization==2.22.0
oslo.utils==3.33.0
pbr==3.1.1
prettytable==0.7.2
pycparser==2.18
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyperclip==1.6.0
python-cinderclient==3.3.0
python-glanceclient==2.9.0
python-keystoneclient==3.14.0
python-novaclient==9.1.1
python-openstackclient==3.13.0
python-openstacksdk==0.5.2
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requestsexceptions==1.3.0
rfc3986==1.1.0
simplejson==3.13.2
six==1.11.0
stevedore==1.28.0
urllib3==1.22
warlock==1.2.0
wrapt==1.10.11
